requirements.txt
mistralai==1.7.0
pydantic>=2.0
python-dotenv